requirements.txt
# Backend API Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
openai==1.3.7
transformers==4.36.0
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
face-recognition==1.3.0
geopy==2.4.0
python-dotenv==1.0.0
stripe==7.8.0
sendgrid==6.11.0
celery-beat==2.5.0
prometheus-client==0.19.0
elastic-apm==6.20.0
sentry-sdk==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0